Fix: Bug in validation of header's offset field
Since da6db173f95ef20aaf4b823bc0cfca4b03875200 the offset field of the
header contains the *absolute* position of the first block. In files
with LZ4 compressed blocks, the header is followed by the jump table
and only then by the blocks. Hence, the header validation was wrong.