Fix Linux / Windows compatibility
For historical reasons MATLAB ships libraries for "small" (32-bit) and
"large" (64-bit) arrays. Both variants are contained in the same
library file.

On Linux, the symbol mxCreateNumericArray is automatically redirected
to mxCreateNumericArray_730, the 64-bit version. On Windows, however,
the mxCreateNumericArray_700 (32-bit variant) is used by default.

To make this library Linux / Windows compatible, we default to 64-bit.