Rust: `Dataset` struct now owns the `root` path
Up to this point, the `Dataset` structure only kept a pointer to the
`root` specified upon construction. In a land purely made out of Rust,
this would be fine due to lifetime guarantees.

But in practice this code is called from, e.g., MATLAB or Python via a
foreign function interface (FFI). These wrappers are not subject to
lifetime checks.

It can thus come to nasty bugs if `root` has a shorter lifetime than the
`Dataset` constructed from it. For an example, see:
https://github.com/scalableminds/webknossos-wrap/issues/13

This commit modifies the `Dataset` structure such that it owns a copy of
`root`. It thus mitigates the problem outlined above.