Bugfix
The Option::map does not modify the value in-place. Because the return
value was not used, this turned into a NOP and left the block index in
an invalid, out-dated state.

This caused trouble when writing partial blocks: First, the original
block content is loaded, modifying the position in the file stream, but
leaving the block index unchanged. Thus, the seek operation before the
second write with the updated block content was turned into a NOP.