Properly handle errors when existing WKW files could not be opened
When reading from a WKW dataset, we treat nonexisting WKW files as
containing only zero-bytes. Previously, we did not separate between

* WKW files not existing and
* existing WKW files that could not be opened.

Both of these cases were treated as if the WKW file did not exist. If an
existing WKW file could not be opened, it was assumed to contain zeros.

Instead, the latter case should be considered and handled as error.